+++ src/main/scala/ESSampleLoadGeneratorApp.scala
@@ -7,7 +7,7 @@ object ESSampleLoadGeneratorApp extends App {
 
   new OptionParser[ESSampleLoadGenerator]("eslg"){
     head("Elasticsearch Sample Load Generator (eslg)", "1.0")
-    opt[Int]('s', "sample-size").optional().action((value, generator) => generator.copy(samples = value))
+    opt[Long]('s', "sample-size").optional().action((value, generator) => generator.copy(samples = value))
     opt[Int]('r', "record-size").optional().action((value, generator) => generator.copy(sampleSize = value))
     opt[String]('n', "cluster-name").required().action((value, generator) => generator.copy(name = value))
     opt[String]('e', "cluster-endpoint").required().action((value, generator) => generator.copy(endpoint = value))
